moving parse/transform out of expressionvisitor
