+++ src/linq/peg/expressionvisitor.ts
@@ -1,5 +1,5 @@
-﻿import ODataParser from './parser/odata-parser'
-import JavascriptParser from './parser/javascript-parser'
+﻿import { ODataParser, JavascriptParser, transform, parse } from './parser'
+
 import { ExpressionStack } from './expressionstack'
 
 import { IExpressionVisitor } from './interface/iexpressionvisitor'
@@ -32,69 +32,27 @@ export class ExpressionVisitor implements IExpressionVisitor {
         return this._expressionStack
     }
 
-    public parseOData(filter: string): IExpression {
-        let ast = ODataParser.parse(filter)
-        try {
-            if(ast) {
-                return this.visit(this.transform(ast))
-            }
-        }
-        catch(ex) {
-            throw new Error(ex.message)
-        }
+    public parse(type: 'odata', predicate: string): ReturnType<typeof parse>
+    public parse(type: 'javascript', lambda: (it: any, ...param: Array<any>) => boolean, ...params: Array<any>): ReturnType<typeof parse>
+    public parse(type: 'javascript', predicate: string): ReturnType<typeof parse>
+    public parse(type: 'odata' | 'javascript', predicate: any, ...params: Array<any>): ReturnType<typeof parse> {
+        let { original, expression } = parse(<any>type, predicate)
 
-        return null
+        return {
+            type, 
+            original,
+            expression: this.visit(expression)
+        }
+    }
+
+    public parseOData(filter: string): IExpression {
+        return this.parse('odata', arguments[0])?.expression
     }
 
     public parseLambda(lambda: string): IExpression
     public parseLambda(lambda: (it: any, ...param: Array<any>) => any): IExpression
     public parseLambda(): IExpression {
-        let lambda: string
-
-        switch(typeof arguments[0]) {
-            case 'string':
-                lambda = arguments[0]
-                break
-
-            case 'function':
-                let regexs = [
-                    /^\(?\s*([^)]*?)\s*\)?\s*(?:=>)+\s*(.*)$/i, //  arrow function; (item) => 5 + 1
-                    /^(?:function\s*)?\(\s*([^)]*?)\s*\)\s*(?:=>)?\s*\{\s*.*?(?:return)\s*(.*?)\;?\s*\}\s*$/i // () => { return 5 + 1 } or function() { return 5 + 1 }
-                ]
-
-                let raw = arguments[0].toString(),
-                    parameters: Array<string>,
-                    expression: string
-
-                for(let regex of regexs) {
-                    let match: RegExpMatchArray
-        
-                    if((match = raw.match(regex)) !== null) {
-                        parameters = match[1].split(',').map((el) => el.trim())
-                        expression = match[2]
-                            .replace(/_this/gi, 'this')
-                            .replace(/\(function\(([^)]+)\)\{return\s(.*?)\}\)/gi, '($1) => $2') // inner arrow functions in methods (for some/every) will probably be converted to a function by babel etc
-        
-                        break
-                    }
-                }
-
-                lambda = `(${parameters.join(', ')}) => ${expression}`
-
-                break
-        }
-
-        let ast = JavascriptParser.parse(lambda)
-        try {
-            if(ast) {
-                return this.visit(this.transform(ast))
-            }
-        }
-        catch(ex) {
-            throw new Error(ex.message)
-        }
-
-        return null
+        return this.parse('javascript', arguments[0])?.expression
     }
 
     public visit(expression: IExpression): IExpression {
@@ -192,247 +150,7 @@ export class ExpressionVisitor implements IExpressionVisitor {
 
         return expression
 
-    }
-
-    /**
-     * transform pegjs expression ast tree to our internal ast tree to make it easier to swap expression parser at a later time
-     * @see
-     * http://www.odata.org/documentation/odata-version-2-0/uri-conventions/
-     *
-     * @param expression pegjs expression object
-     * @returns 
-     */
-    private transform(expression: Record<string, any>): IExpression {
-        let child: IExpression
-
-        switch(expression.type) {
-            case 'LambdaExpression':
-                return new LambdaExpression(Array.isArray(expression.arguments) ? expression.arguments.map((arg) => this.transform(arg)) : [], this.transform(expression.expression))
-
-            case 'Identifier':
-                return new IdentifierExpression(expression.name)
-
-            case 'MemberExpression':
-                switch(expression.property.type) {
-                    case 'CallExpression':
-                        child = this.transform(expression.property);
-                        (<MethodExpression>child).caller = this.transform(expression.object)
-
-                        return child
-
-                    case 'MemberExpression':
-                        let current = expression,
-                            stack = []
-
-                        do {
-                            stack.push(this.transform(current.object))
-
-                            if(current.property.type != 'MemberExpression')
-                                break
-
-                            current = current.property
-                        } 
-                        while (true)
-
-                        if(current.property.type == 'CallExpression') {
-                            // we want stacked memberexpressions as caller for handyness
-                            child = this.transform(current.property);
-
-                            (<MethodExpression>child).caller = stack.reduceRight((caller, member) => {
-                                if(!caller)
-                                    return member
-                                else 
-                                    return new MemberExpression(member, caller)
-                            }, null)
-
-                            return child
-                        }
-
-                        // fall through
-
-                    default:
-                        return new MemberExpression(this.transform(expression.object), this.transform(expression.property))
-                }
-                
-            case 'CallExpression':
-                switch(expression.object.type) {
-                    case 'Identifier':
-                        return new MethodExpression(expression.object.name, Array.isArray(expression.arguments) ? expression.arguments.map((arg) => this.transform(arg)) : [], null)
-
-                    default:
-                        throw new Error('Caller of method expression is not a Identifier, but is ' + expression.object.type)
-                }
-
-            case 'DateTimeLiteral':
-                return new LiteralExpression(new Date(expression.value))
-
-            case 'DateLiteral':
-                let value = new Date(expression.value)
-                
-                return new LiteralExpression(new Date(Date.UTC(value.getUTCFullYear(), value.getUTCMonth(), value.getUTCDate(), 0, 0, 0, 0)))
-
-            case 'NumberLiteral':
-                return new LiteralExpression(Number(expression.value))
-
-            case 'BooleanLiteral':
-                return new LiteralExpression(expression.value == true || expression.value == 'true' ? true : false)
-
-            case 'NullLiteral':
-                return new LiteralExpression(null)
-
-            case 'Literal':
-                return new LiteralExpression(expression.value)
-
-            case 'ConditionalExpression':
-                return new ConditionalExpression(this.transform(expression.test), this.transform(expression.left), this.transform(expression.right))
-
-            case 'ObjectLiteral':
-                return new ObjectExpression(Array.isArray(expression.properties) ? 
-                    expression.properties.map(({ key, value }) => {
-                        switch(key.type) {
-                            case 'Identifier':
-                                key = {
-                                    type: 'Literal',
-                                    value: key.name
-                                }
-                                break
-                            default:
-                                break
-                        }
-                        return { key: this.transform(key), value: this.transform(value) }
-                    }) : [])
-
-            case 'TemplateLiteral':
-                if(Array.isArray(expression.values) && expression.values.length > 0) {
-                    let literals: Array<ILiteralExpression> = [], 
-                        expressions: Array<IExpression> = [],
-                        first = expression.values[0]
-
-                    if(first.type == 'TemplateExpression') 
-                        literals.push(<ILiteralExpression>this.transform({ type: 'Literal', value: '' }))
-
-                    literals.push(...expression.values
-                        .filter(value => value.type == 'Literal')
-                        .map(value => <ILiteralExpression>this.transform(value))
-                        .map(literal => { literal.value = literal.value.replace(/(\\)?\$(\\)?\{/g, '${'); return literal })
-                    )
-                    expressions.push(...expression.values.filter(value => value.type == 'TemplateExpression').map(value => this.transform(value.value)))
-
-                    return new TemplateLiteralExpression(literals, expressions)
-                }
-
-                return new TemplateLiteralExpression([], [])
-
-            case 'ArrayExpression':
-                return new IndexExpression(this.transform(expression.object), this.transform(expression.index))
-
-            case 'ArrayLiteral':
-                return new ArrayExpression(Array.isArray(expression.elements) ? expression.elements.map((arg) => this.transform(arg)) : [])
-
-            case 'LogicalExpression':
-            case 'LogicalExpression':
-                switch(expression.operator) {
-                    case '&&':
-                        return new LogicalExpression(LogicalOperatorType.And, this.transform(expression.left), this.transform(expression.right))
-                    case '||':
-                        return new LogicalExpression(LogicalOperatorType.Or, this.transform(expression.left), this.transform(expression.right))
-                }
-                break
-
-            case 'RelationalExpression':
-                switch(expression.operator) {
-                    case '==': // equal
-                        return new LogicalExpression(LogicalOperatorType.Equal, this.transform(expression.left), this.transform(expression.right))
-
-                    case '!=': // not equal
-                        return new LogicalExpression(LogicalOperatorType.NotEqual, this.transform(expression.left), this.transform(expression.right))
-
-                    case '<': // lesser
-                        return new LogicalExpression(LogicalOperatorType.Lesser, this.transform(expression.left), this.transform(expression.right))
-
-                    case '<=': // lesser or equal
-                        return new LogicalExpression(LogicalOperatorType.LesserOrEqual, this.transform(expression.left), this.transform(expression.right))
-
-                    case '>': // greater
-                        return new LogicalExpression(LogicalOperatorType.Greater, this.transform(expression.left), this.transform(expression.right))
-
-                    case '>=': // greater or equal
-                        return new LogicalExpression(LogicalOperatorType.GreaterOrEqual, this.transform(expression.left), this.transform(expression.right))
-                }
-                break
-
-            case 'PostfixExpression':
-                switch(expression.operator) {
-                    case '--':
-                        return new UnaryExpression(UnaryOperatorType.Decrement, UnaryAffixType.Postfix, this.transform(expression.argument))
-                    case '++':
-                        return new UnaryExpression(UnaryOperatorType.Increment, UnaryAffixType.Postfix, this.transform(expression.argument))
-                }
-                break
-
-            case 'UnaryExpression':
-                switch(expression.operator) {
-                    case '!':
-                        return new UnaryExpression(UnaryOperatorType.Invert, UnaryAffixType.Prefix, this.transform(expression.argument))
-                    case '~':
-                        return new UnaryExpression(UnaryOperatorType.Complement, UnaryAffixType.Prefix, this.transform(expression.argument))
-                    case '+':
-                        return new UnaryExpression(UnaryOperatorType.Positive, UnaryAffixType.Prefix, this.transform(expression.argument))
-                    case '-':
-                        return new UnaryExpression(UnaryOperatorType.Negative, UnaryAffixType.Prefix, this.transform(expression.argument))
-                    case '--':
-                        return new UnaryExpression(UnaryOperatorType.Decrement, UnaryAffixType.Prefix, this.transform(expression.argument))
-                    case '++':
-                        return new UnaryExpression(UnaryOperatorType.Increment, UnaryAffixType.Prefix, this.transform(expression.argument))
-                }
-                break
-
-            case 'ShiftExpression':
-                switch(expression.operator) {
-                    case '<<':
-                        return new BinaryExpression(BinaryOperatorType.LeftShift, this.transform(expression.left), this.transform(expression.right))
-                    case '>>':
-                        return new BinaryExpression(BinaryOperatorType.RightShift, this.transform(expression.left), this.transform(expression.right))
-                    case '>>>': // zero-fill right-shift 
-                        return new BinaryExpression(BinaryOperatorType.RightShift, this.transform(expression.left), this.transform(expression.right))
-                }
-                break
-
-            case 'BitwiseExpression':
-                switch(expression.operator) {
-                    case '|':
-                        return new BinaryExpression(BinaryOperatorType.Or, this.transform(expression.left), this.transform(expression.right))
-                    case '^':
-                        return new BinaryExpression(BinaryOperatorType.ExclusiveOr, this.transform(expression.left), this.transform(expression.right))
-                    case '&':
-                        return new BinaryExpression(BinaryOperatorType.And, this.transform(expression.left), this.transform(expression.right))
-                }
-                break
-
-            case 'BinaryExpression':
-                switch(expression.operator) {
-                    case '+': // addition
-                        return new BinaryExpression(BinaryOperatorType.Addition, this.transform(expression.left), this.transform(expression.right))
-
-                    case '-': // subtraction
-                        return new BinaryExpression(BinaryOperatorType.Subtraction, this.transform(expression.left), this.transform(expression.right))
-
-                    case '*': // multiplication
-                        return new BinaryExpression(BinaryOperatorType.Multiplication, this.transform(expression.left), this.transform(expression.right))
-
-                    case '/': // division
-                        return new BinaryExpression(BinaryOperatorType.Division, this.transform(expression.left), this.transform(expression.right))
-
-                    case '%': // modulus
-                        return new BinaryExpression(BinaryOperatorType.Modulus, this.transform(expression.left), this.transform(expression.right))
-                }
-                break                
-        }
-
-        throw new Error('Expression type "' + expression.type + '" is unknown')
-    }
-
-  
+    }  
 }
 
 //export { IExpression, Expression, ExpressionType } from './expression';
